+++ exoskeleton/__init__.py
@@ -2,9 +2,10 @@
 
 import importlib.metadata
 
+from exoskeleton.action_types import ActionType
 from exoskeleton.core import Exoskeleton
 
-__all__ = ['Exoskeleton']
+__all__ = ['ActionType', 'Exoskeleton']
 
 NAME = "exoskeleton"
 __version__ = importlib.metadata.version("exoskeleton")